working version, with Makefile for lubuntu
