+++ Makefile
@@ -1,3 +1,6 @@
+mkfile_path = $(abspath $(lastword $(MAKEFILE_LIST)))
+current_dir := $(notdir $(patsubst %/,%,$(dir $(mkfile_path))))
+
 all:
 	$(MAKE) -C dependencies/pgn-extract
 	mv dependencies/pgn-extract/pgn-extract  bin/pgn-extract
@@ -5,12 +8,27 @@ all:
 	$(MAKE) -C dependencies/uci-analyser
 	mv dependencies/uci-analyser/analyse bin/analyse
 
-	g++ -std=c++17 apgn.cpp -o apgn.out
-	
+	chmod +x bin/engines/stockfish10_linux
+
+# g++ -std=c++17 apgn.cpp -o apgn -DDEBUG -fsanitize=address -g
+	g++ -std=c++17 apgn.cpp -o apgn -O3
+
 test:
-	./apgn.out "./test/first.pgn" "./test/myout.pgn" W
+	./apgn ./pgn_samples/first.pgn W
+
+test_clean:
+	rm ./pgn_samples/firstAnalyzed.pgn
+
+install:
+	ln -s $(dir $(abspath $(lastword $(MAKEFILE_LIST))))apgn /usr/bin
+
+uninstall:
+	rm /usr/bin/apgn
+	@rm ./bin/analyse ./bin/pgn-extract ./apgn
 
 clean:
-	$(MAKE) -C dependencies/pgn-extract clean
-	$(MAKE) -C dependencies/uci-analyser clean
-	rm execute.out ./bin/analyse ./bin/pgn-extract
+	@echo "removing pgn-extract object files"
+	@$(MAKE) -C dependencies/pgn-extract clean
+	@echo "removing uci-analyse object files"
+	@$(MAKE) -C dependencies/uci-analyser clean
+	@echo "removing analyse-pgn object files"
