--- env.cpp
@@ -1,17 +1,0 @@
-#include <iostream>
-#include "envgrabber.hpp"
-
-int main()
-{
-    std::vector<std::string> env_variables = apgn_env::grab("env.txt",{
-        "CHESS_ENGINE",
-        "WORKER_THREADS",
-        "ANALYSIS_DEPTH",
-        "OPENING_MOVE_TO_SKIP"
-    });
-
-    for(auto e: env_variables)
-        std::cout<<e<<"\n";
-
-    return 0;
-}
