+++ env.txt
@@ -1,4 +1,36 @@
+You can change the values of this files to your own liking
+
+----------------------------------------------------------------------------
+
 CHESS_ENGINE=stockfish10_linux
-WORKER_THREADS=3
-ANALYSIS_DEPTH=18
+
+this is the chess engine executable that the program will use to analyse
+chess pgn games, if you want to use another engine, put the chess
+engine's binary executable or .exe file inside the 'bin/engines'
+folder and set CHESS_ENGINE's value into the filename of the new
+chess engine you want to use, by default it uses the 'stockfish10_linux'
+
+----------------------------------------------------------------------------
+
+WORKER_THREADS=1
+
+this is the number of threads the chess engine will use, bigger
+number means, faster analysis of .pgn files, as long as your cpu
+can allocate that number of threads, if you exceed your cpu's limit...
+it might slow down your analysis, so find out your cpu's maximum threads
+first before setting this to a very large number, usually moder computers
+have maximum total of 4 threads
+
+----------------------------------------------------------------------------
+
+ANALYSIS_DEPTH=12
+
+the larger the value, the stronger the analysis, but larger values also
+means longer analysis time
+
+----------------------------------------------------------------------------
+
 OPENING_MOVE_TO_SKIP=6
+
+note, this value refers to the moves itself not the turns, example
+pawn to e4, then pawn to e5 is the first turn, and 2 moves
